requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
Pillow==6.0.0
praw==6.2.0
prawcore==1.0.1
requests==2.21.0
six==1.12.0
update-checker==0.16
urllib3==1.24.3
websocket-client==0.56.0
